requirements.txt
pytest-asyncio
httpx

fastapi
uvicorn
beanie
motor
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
jinja2
aiofiles
mutagen
email-validator